+++ R/dearseq_package.R
@@ -8,8 +8,8 @@
 #'\tabular{ll}{
 #'Package: \tab dearseq\cr
 #'Type: \tab Package\cr
-#'Version: \tab 1.3.0\cr
-#'Date: \tab 2021-03-09\cr
+#'Version: \tab 1.4.0\cr
+#'Date: \tab 2021-08-31\cr
 #'License:\tab \href{http://www.gnu.org/licenses/gpl-2.0.txt}{GPL-2}\cr
 #'}
 #'The two main functions of the \code{dearseq} package are